Dump bridge pool assignments to a file for statistics.
With this patch we dump the list of currently running bridges with
information about their assigned rings, subrings, and file buckets to a
local file.  The idea is to publish sanitized versions of these assignment
files and analyze how the assignment affects a bridge's usage.

The assignment file is written on startup and after receiving a HUP signal
and parsing new bridge descriptors.  Note that the assignments file is not
updated when bridges are dumped to file buckets; in that case the changed
assignments to file buckets will be reflected in the assignments file
after the next HUP.

Also note that the assignment file only contains bridges that are believed
to be running from parsing the last network status.  As a result, bridges
that are contained in file buckets, but that are not believed to be
running, won't be contained in the assignment file.