Refactor bridgedb.Main.run() and bridgedb.Main._reloadFn().
The main entry point and server reload function are refactored to recieve the
parsed options from the commandline script, scripts/bridgedb.

It is necessary to pass *args to `_reloadFn()` (as well as its original
stubbed lambda function), because the arguments are technically unknown to the
scope of the `reload()` function embedded in `bridgedb.Main.startup()` when it
is evaluated at compile time (they are not known until later when
`bridgedb.Main.run()` is called via scripts/bridgedb).