Start explaining the modification to remove Oblivious Transfer.
Also add detailed, high-level description of the original usage of OT for
bridge distribution, and why this is not needed within the modified threat
model. However, it occurred to me while writing out the details that the k-TAA
blind signatures (Au, Susilo, et al.) used in the original scheme *also*
require a bilinear pairing, and are used independently to the usage of OT for
creating the zero-knowledge PoKs of valid blind signatures on user commitments
to the values representing the user's numbers of Credits (the coin-like things
generated through one's bridges not getting blocked) and the user's timestamp
for the last request for an Invite Ticket.

As such, the signature scheme must either be re-evaluated and a different
scheme used, or else we likely do not save any implementation overhead by
ripping out the Oblivious Transfer (although doing so *does* still save on the
rather high computational and network bandwidth complexities involved in OT).