Move Main.Conf → persistent.Conf.
In order to store `Conf` config objects persistently, the Conf class must be
accessible in the bridgedb.persistent module. However, the bridgedb.persistent
module must be importable into bridgedb.Main in order to use it. This creates
a circular import dependency. To fix it, I moved Main.Conf to persistent.Conf.

Also, there were bugs in the Conf class which caused Conf instances to store
__builtin__ classes. To fix this, I changed:

     for key, value in attrs.items():
         if key.upper() == key:
             self.__dict__[key] = value

to

     for key, value in attrs.items():
         if key == key.upper():
             if not key.startswith('__'):
                 self.__dict__[key] = value

 * FIXES a bug in Conf class causing Python's __builtin__ classes to be
   reloaded into the global scope.