Rewrite Main.loadConfig() to only log after the first run.
There is a bit of a chicken-egg problem here, in that we need to parse our
config file before we can configure logging, but we should start logging
before changing settings, modifying the config file settings, and creating out
persistent.State storage. To get around this, loadConfig() is now modified to
only make calls to the logging module the second time the function is called
(and all the times after).

 * MOVE all config/setting changing code from Main.startup() to
   Main.loadConfig() so that it is reparsed on SIGHUP correctly.