Change setup.py/requirements.txt to allow URI/VCS resources.
The maintainer of ipaddr has ceased uploading to PyPI, and so to get pip to
download it, we need to add its URI and hash to requirements.txt. The
get_requirements function in setup.py must be refactored to handle parse these
URIs correctly, and return dependency_links in addition to install_requires.