Fix a bug where relative rundir/config options didn't work with subCommands.
Because subclasses of bridgedb.parse.options.BaseOptions will call their own
parseOptions() method, which calls their own postOptions() as well as all
parent class postOptions() methods, we have to set the absolute paths of the
config file and the runtime directory as globals of the bridgedb.parse.options
module. Otherwise, they will get reset in the parent classes, who do not have
the instance variables of the subclass which did the original commandline
parsing, and thus the config file and runtime directory get reset to their
default settings by the parent classes. This then raises errors when an
attempt to read the config file is made in bridgedb.Main.loadConfig(), because
the file is (likely) not found.

 * FIXES a bug introduced by the changes for #9872, pertaining to setting the
   runtime directory and config file paths when using options.subCommand.