Catch all parser exceptions in parse.networkstatus.parseRLine().
Rather than raising exceptions during parsing, we should try to log
appropriate messages. If the error had a known cause, i.e. one that is due to
a badly formatted descriptor, then we should make note of the relevant
information from the bad descriptor, and reset any affected variables to
None. If the error is possibly do to a newly implemented change in the
descriptor format in tor, we should likewise log the event, generate a loud
warning about a possible descriptor format change, and continue as before.

In networkstatus documents and [bridge-]server-descriptors, it is possible
that the `ORAddress` lines may soon change:
    1. to allow more than one ORAddress line, and
    2. to allow extra IPv4 addresses in the ORAddress lines.

 * CHANGE bridgedb.parse.networkstatus.parseRLine() function to catch and
   handle any and all InvalidNetworkstatusDescriptorDigest
   InvalidNetworkstatusRouterIdentity, and NetworkstatusParsingErrors,
   exceptions raised during parsing.