Catch all parser exceptions in parse.networkstatus.parseRLine().
Rather than raising exceptions during parsing, we should try to log
appropriate messages. If the error had a known cause, i.e. one that is
due to a badly formatted descriptor, then we should make note of the
relevant information from the bad descriptor, and reset any affected
variables to None.

If the error is possibly do to a newly implemented change in the
descriptor format in tor, we should likewise log the event, generate a
loud warning about a possible descriptor format change, and continue as
before. In networkstatus documents and [bridge-]server-descriptors, it
is possible that the `ORAddress` lines may soon change (#9729):

  1. to allow more than one ORAddress line, and
  2. to allow extra IPv4 addresses in the ORAddress lines.

Additionally, remove exception and handlers which are never raised in
parseRLine().

  1. The catch for `IndexError`s in networkstatus.parseRLine() is
  removed:

    At the beginning of `parseRLine()`, a `NetworkstatusParsingError` is
    raised if there are less than eight fields in the descriptor, so
    there is no way that an `IndexError` would ever get raised. Coverage
    branch reports also show that this `except:` block is never touched.

  2. Remove the `InvalidNetworkstatusDescriptorDigest` exception class:

    …as well as all corresponding code for raising and handling it. The
    way it was used, there is no possible way that these lines will get
    executed, because the descriptor digest will either be a str or
    None. If None, then an error is raised earlier for having too few
    fields in the 'r'-line. If a str, then `binascii.a2b_base64() will
    always produce some kind of string, and its exceptions are handled
    separate. Therefore, there is no way for the descriptor digest to be
    0, None, or False.

In `bridgedb.parse.networkstatus.parseRLine()` function:

 * CHANGE to catch and handle all `InvalidNetworkstatusRouterIdentity`,
   and `NetworkstatusParsingErrors`, exceptions raised during parsing.
 * REMOVE unnecessary `except IndexError` block.
 * REMOVE `InvalidNetworkstatusDescriptorDigest` exception class.
 * CHANGE logging of caught exceptions to use `exc`, not
   `exc.message`. The later has been deprecated since Python2.6.