Accessing Exception.message has been deprecated since Python2.6.
I'm was little suprised to see the DeprecationWarning messages saying this,
but I suppose I usually kill the display of warnings in most of the things I
work on since they can come from any of the pulled-in dependency chains and I
dislike the noise. Well, better late than never.