Refactor networkstatus.parseALine() to only return portlist iff ip.
There were some cases raised with the unittests where
networkstatus.parseALine() would return an addr.PortList when the IP
address had failed to parse. Having a PortList without an IP address is
useless, so instead we should deal with parsing the IP first, and
checking its validity. Then, if we have the IP, we should parse and
check the portlist.