Add test_Tests.TrialAdaptedDoctests class.
This class uses the dynamic doctest discovery mechanism,
`test_Tests.generateTrialAdaptedDoctestSuite`, to run all discovered
doctests. The generated methodnames in this `TestCase` correspond to
location in the source code where the doctest was discovered.

It also (somewhat) magically manages to print exception/failure
tracebacks for failed doctests in a manner which is compatible with
twisted.trial. The printed tracebacks will tell you:

  1. Where the doctest is, and which line caused the exception.

  2. What the expected result was supposed to be (according to the
     doctest), compared with what it actually was when the doctest code
     was run.