Add example extrainfo `transport` line for scramblesuit.
Philipp Winter's scramblesuit paper has a section on creating easily
handwritable shared secrets. [0]

To my memory, this excludes using the character `O` (capital o) and the
integer `1`, because they are easily confused for the integer '0' and
the character 'l', respectively. However, Philipp mentioned on IRC that
scramblesuit `password=` fields in the `transport` line are actually
just the set of base32 characters. [1]

 * ADD a line to doc/DESCRIPTORS.md which shows an example
   scramblesuit-enabled bridge which uses a shared secret passed through
   BridgeDB. See #9013.

[0]: http://arxiv.org/abs/1305.3199
[1]: https://tools.ietf.org/html/rfc4648