Use SSL when requesting CAPCHAs from reCaptcha.
This sounds bad. It is. What's worse: this only fixes half the
problem. The recaptcha.client.captcha.submit() function in the reCaptcha
API [0] is hardcoded to submit the server's (in this case, BridgeDB's)
private API key, the client's IP address (BridgeDB sends a faked one),
and the client's solution to the CAPTCHA, all over HTTP, in glorious
plaintext. Hooray.

[0]: https://recaptcha.googlecode.com/svn/trunk/recaptcha-plugins/python/recaptcha/client/captcha.py