Add a mocked DummyBridge class to test_HTTPServer.py.
This is used to test that bridges are actually returned on
'bridges.html' results pages, and is missing all of the real class's
methods except for `getConfigLine()`. The latter method needs to be
mocked because is used in `HTTPServer.WebResourceBridges.render()` to
get the bridge line to display to a client.