Avoid using RSA in GimpCaptcha
Totally untested, and as far as I'm concerned, this spaghetti monster is
totally untestable.  But if it works at all, it'll be just as secure as the
old code, and *much* faster.

Also adds an experimental folding method, and logs which kinds of folding
(if any) were needed to match a response.  (I suspect that users can handle
case sensitivity for all but a few letters; let's find out.)