Kill bridgedb.I18n and add new bridgedb.strings module.
This new module holds translated strings which are used in several places in
the code, in order to save our helpful translators from having to translate
the same boring string over and over again.

This has been redesigned to permit these strings to be formatted for both the
email and HTTPS distributors.