Add XXX note about sending BridgeRequest object to Distributor.
If we assume that the Distributor can be run separately, then we should
assume that the email servers will send their EmailBasedDistributor a
`bridgedb.bridgerequest.BridgeRequest` object describing the incoming
request. It should be the Distributor's job to transform this request,
if necessary, into something which can talk to the DatabaseManager (see
comment:1 on #12029). [0] That is, the Distributor should be capable of
transforming BridgeRequests it receives into JSON.

[0]: https://trac.torproject.org/projects/tor/ticket/12029#comment:1