Remove `answerParameters` from Hashring to Distributor __init__() method.
Before, when a `Distributor` inherited from a `BridgeHolder` (which was
a crappy "base class" for both hashrings and distributors), the
`BridgeHolder` base class expected to receive a `answerParameters`
arguments. This argument is only used by a `Distributor`, because an
`answerParameters` is meant to be an object which describes what types
of Bridges the Distributor should give to a client.

 * ADD `answerParameters` argument to `bridgedb.distribute.Distributor`,
   and set up the `Distributor`'s hashring based upon the
   `answerParameters`.

 * REMOVE `answerParameters` argument from `bridgedb.hashring.Hashring`.