Fix test suite for bridgedb.scheduled.Unscheduled
Previously, the test suite checked only that Unscheduled.getInterval would
return a particular value when invoked with no argument.  Now, the test
suite checks the only property of Unscheduled.getInterval that its callers
rely on: that it returns the same string regardless of which time is passed
in to it.