Revert check for SMTP/email header canonical hostname equivalence.
Signed-off-by: Isis Lovecruft <isis@torproject.org>

For now, we need to revert this check to get the email distributor to
function. We should look into this issue in order to get BridgeDB in a
state where instances of it are runnable by other organisations to hand
out their own bridges. [OTHER_ORG]

Fixing this is essential for #12086.