Be more careful about how we split bridge lines in test_https.py.
This is done to ensure that if we don't get the bridges.html page in
response (i.e. if the CAPTCHA solution was wrong, or something is just
plain wrong with the webserver that is supposed to be running in the
background), that the BeautifulSoup HTML tag handling in test_https.py
will be a bit more robust.