Workaround strange file descriptor issue in Stem.
For reasons I am unable as of yet to determine, if I call
`stem.descriptor.router_status_entry._parse_file()` with its `start_position`
parameter, giving it the byte offset in the file to start reading at, as well
as the file descriptor itself, that Stem function does *not* reliably `seek()`
to that byte offset in the file.

The failure results of the last commit, e47247869, can be seen here:
https://travis-ci.org/isislovecruft/bridgedb/jobs/33874252

The fix -- don't ask me why it works, although I suspect either the scoping of
Python's `with` syntax or some issue with passing file descriptors from an
asynchronous program to a blocking one -- is to call `seek()` before passing
the file descriptor.