Add error handling for cases when Stem cannot parse an extrainfo file.
If a file containing bridge-extrainfo descriptors cannot be parsed the
first time, and validation was enabled, then we parse it again with
Stem's validation functionality disabled. If we still can't parse the
file, we save a timestamped copy of the file for later debugging an
continue with the next file.