Bridge._constructBridgeLine() won't be handed an invalid addresses.
The only way that Bridge._constructBridgeLine() is called is through
Bridge.getBridgeLine() when vanilla bridges are requested.  Because of
all the sanitisation in Stem and the rest of BridgeDB's parsers, no type
of bridgerequest, no matter how malformed, will ever result in an
invalid addrport from Bridge._getVanillaForRequest().

Secondly, we'll never get anything other than IPv4 and IPv6 addresses,
for the same reasons.

 * REMOVE lines which raise MalformedBridgeInfo in
   Bridge._constructBridgeLine().