Change Main.run() to take a reactor as a parameter.
This makes unittesting easier, as we can then pass mocked reactors with
less risk of touching the real one. (However, to pass mocked reactors,
we'd probably need to check the code for the EmailDistributor and
IPBasedDistributor in bridgedb.Dist to ensure that neither are touching
the reactor either.)

This also changes the main loop to only call reactor.run() if the
reactor is not already running.