Fix bug in proxy.loadProxiesFromFile().
Checking `if proxySet` when the ProxySet was empty (as at server
startup, particularly when loading the PROXY_LIST_FILES), would evaluate
to False, cause the proxies from the files *not* to be loaded into the
empty ProxySet.

This did not effect loading the list of Tor Exit relays, since we use
scripts/get-tor-exits and load the list into memory while calling
ProxySet.add() for each one (rather than loading Tor Exit relays from a
downloaded file, like we used to do).