Add HTTPS_ROTATION_PERIOD and EMAIL_ROTATION_PERIOD config options.
This implements configuration of the hashring rotation periods for
BridgeDB's distributors via the configuration file.

The behaviour for hashring rotation for the EmailBasedDistributor is
such that, when `bridgedb.email.autoresponder.createResponseBody()`
calls `EmailBasedDistributor.getBridgesForEmail(EMAIL_ADDRESS, EPOCH)`
with the `EPOCH` set to the start of the current EMAIL_ROTATION_PERIOD,
the client's position in the hashring is determined by the HMAC of the
string `"<EPOCH>EMAIL_ADDRESS"`.  Therefore, the EMAIL_ROTATION_PERIOD
directly effects where the client is placed in the hashring, resulting
in different bridges for the user, depending on whether the period has
elapsed.  With the default setting of `"1 day"`, and taking into account
also that the EmailBasedDistributor only responds to a particular user
once per three hours, this results in the client being able to ask for
(and receive) vanilla bridges (starting from hashring position A) at
9:00, obfs3 bridges (also from position A) at 12:00, obfs4 bridges (from
position A again) at 15:00, and so on, and finally different vanilla
bridges (from hashring position B) the next morning at 9:00.

For the IPBasedDistributor, the behaviour for hashring rotation is that
the client's hashring position is determined by the HMAC of the string
`"<EPOCH>AREA"` where `EPOCH` is again the start of the current
HTTPS_ROTATION_PERIOD, and the `AREA` is the `/16` subnet which the
client's IP address resides within.  If the client is using Tor or some
other open proxy, then the client's hashring position is determined by
`"known-proxy<EPOCH>GROUP"` where `EPOCH` is the same as before, and
`GROUP` is a number (currently 1 through 4, inclusive) deterministically
derived from the IP address of the Tor Exit relay or open proxy that the
client is using.  Using `GROUP` causes there to only be 4 sets of
bridges available to any and all Tor/proxy users at a given time.  Hence
additionally using `EPOCH` rotates the set of 4 bridges available.  The
default setting is `"3 hours"`, causing all Tor/proxy users to have 4
different sets of bridges every three hours, while non-Tor users have a
new set of bridges (probably) unique to their IP address available every
three hours.

These defaults for HTTPS_ROTATION_PERIOD and EMAIL_ROTATION_PERIOD will
likely need to be changed and fiddled with to make the behaviour an
optimum balance between user-friendly and resilient to enumeration.

 * FIXES #1839: https://bugs.torproject.org/1839.
   This implement Roger's suggestion to add the rotation periods as
   configuration file options.