Why were there ever unreachable subhashrings? Kill them with fire.
As part of #4297, $SOMEONE added the prepopulateRings() methods, which
are used when BridgeDB parses incoming descriptors.  These methods add
subhashrings and allocate bridges to them.  When a client requests
bridges, they are either requesting IPv4 or IPv6 bridges, and thus
clients *always* have either filterBridgesByIPv4 or filterBridgesByIPv6,
respectively.  So why is there an extra subhashring with no IP version
filters if clients *can never get to it*?

Blame $SOMEONE.  The $SOMEONE who authored the commit below would be a
good start.

 * BUGFIX on e6ce57e728802689544c130867edcadfeecd38ec.
 * REMOVE the "filterless" subhashring from the IPBasedDistributor.
 * UPDATE documentation for IPBasedDistributor.prepopulateRings().