Make the @port property common to both PluggableTransports and Bridges.
Because it seems awkward to require doing type checks in order to know
which port a bridge-like thing is listening on.

That is, previously, we could access PluggableTransport.port and
Bridge.orPort, but *not* Bridge.port.  Moving the port @property from
the PluggableTransport class to the BridgeAddressBase class (and
updating the Bridge.orPort @property to point at the port @property)
allows us to access the port property of either one without caring which
type it is.

In addition to allowing us to forego type checks, this allow allows us to
define an interface which essentially says that "a bridge-like thing is a
thing which has a fingerprint, address, and a port", which again allows us to
generalise code to deal with "bridge-like" things, rather than dealing with a
tuple of classes which happen to be considered "bridge-like" at the time of
writing.

 * MOVE the port @property from the PluggableTransport class to the
   BridgeAddressBase class.

 * CHANGE the Bridge.orPort @property to point at the new port
   @property.